+++ R/methods_aod.R
@@ -3,7 +3,7 @@
 get_predict.glimML <- function(model,
                                newdata = insight::get_data(model),
                                vcov = FALSE,
-                               conf.level = 0.95,
+                               conf_level = 0.95,
                                type = "response",
                                ...) {