+++ R/methods_bife.R
@@ -5,7 +5,7 @@
 get_predict.bife <- function(model,
                              newdata = insight::get_data(model),
                              vcov = FALSE,
-                             conf.level = 0.95,
+                             conf_level = 0.95,
                              type = "response",
                              ...) {