+++ R/methods_brms.R
@@ -25,7 +25,7 @@ get_coef.brmsfit <- function(model, ...) {
 get_predict.brmsfit <- function(model,
                                 newdata = insight::get_data(model),
                                 vcov = FALSE,
-                                conf.level = 0.95,
+                                conf_level = 0.95,
                                 type = "response",
                                 ...) {