+++ R/methods_quantreg.R
@@ -3,7 +3,7 @@
 get_predict.rq <- function(model,
                            newdata = insight::get_data(model),
                            vcov = NULL,
-                           conf.level = 0.95,
+                           conf_level = 0.95,
                            type = NULL,
                            ...) {